<refactor> Add additional options to PR template
Add a refactor option as a significant number of the PRs generated fall
into this category.

Add an item to the checklist to confirm that none of the standard
items are applicable.

Rename "Breaking Actions" to "Followup actions" so the item can capture
basically anything that should be done once the PR is merged.