<refactor> Rationalise script usage
As the automation and generation scripts are now in one tree, use the
same instance of utility.sh and contextTree.sh to avoid needing to keep
both in sync.

As an interim measure, include symlinks in the old location for any
explicit references in jenkins jobs. These should be changed to
reference the true location, and at some point the symlinks will be
removed, hence the TODO notice.