<feature> Warn on invalid build references
The implementation of accept and promote processing means that an
invalid build reference will prevent successful processing. During the
addition of new units, it is sometimes desirable to ignore any units
that don't reference a valid image.

The default behaviour is to trigger a fatal error but this change
permits a WARN_ON_INVALID_BUILD_REFERENCES variable to be set to "true"
which will then cause warnings to be displayed for invalid units, and
for the units then to be skipped from a processing perspective.