feat: Re-encrypt and list cmk crypto operations
Add the ability to re-encrypt existing ciphertext. This makes creating
new segments much easier where files or attributes need to have the same
plaintext values. Instead of having to obtain the plaintext, existing
ciphertext can be copied and then re-encrypted.

The re-encrypt command also provides the original key id as an output,
so this has been used to provide the ability to list the cmk used to
encrypt any ciphertext. The resulting value is then used to query what
alias the key has.