feat: Bare workflow support
SDK37 deprecated the use of Expokit in favour of the bare workflow.

Add support for bare workflow configurations.

Add support for app extensions which have their own Info.plist files,
bundle ids etc.

Use the APP_REFERENCE, if present, to define the ShortVersionString.
This needs to be set in app extensions to the same value as the app
bundle.

Ensure the metro bundler is shutdown after each build. The workspace
contents get deleted each time so it doesn't runproperly if left
running.