feat: Add content encoding metadata
If a filename contains the string "-precompressed", assume it has been
gzip compressed and so when copying it to S3, set the necessary metadata
to serve it correctly.

The file should still bear the correct extension so that the
Content-Type is set correctly during S3 upload.

The initial use case for this is to provide compression in cloudfront for
files greater than 10Mb, which cloudfront won't dynamically compress.

A lambda@edge function can detect the "-precompressed" in the requested
URL and change the origin URL to target the precompressed file.