fix: Shared account/product/environment/segment names
Add more robust checking for account and product directories to allow
for the possibility of them sharing the same name e.g.
product=environment or account=product.

Because the logic supports the pre and post v2.0 cmdb formats, also a
range of cmdb granularities, its possible there are still configurations
were the new logic will fail, but we will deal with them when they
happen. For now, this will address the common examples identified above.