feat: Branch based deployment plans
Add the ability to generate the templates and change sets that would
apply should a deploy be performed.

This is the first step in making the idea of "preparing" releases more
temporary rather than being a mainstream part of the deployment
process.

A deployment plan is a separate branch that captures what templates and
change sets would be used based on the current state of the
infrastructure.

The branch can be deleted at any time without affecting the ability to
perform future deploys.