feat: Branch based deployment plans (#70)
* feat: Branch based deployment plans

Add the ability to generate the templates and change sets that would
apply should a deploy be performed.

This is the first step in making the idea of "preparing" releases more
temporary rather than being a mainstream part of the deployment
process.

A deployment plan is a separate branch that captures what templates and
change sets would be used based on the current state of the
infrastructure.

The branch can be deleted at any time without affecting the ability to
perform future deploys.

* fix: Dryrun handling

* fix: Correct repo management for branches

* feat: Plans can apply to any level

* refactor: Treat a plan as a dryrun of manageUnits

* fix: Include deployment mode in plan branch name

If multiple deployment modes are being used, e.g. for blue/green
switchovers, it should be possible to see the plan for each
independently. Thus the deployment mode, if present, isincluded in the
plan branch name.

* fix: Back out addition of DEPLOYMENT_MODE_PLAN

Plans are now implemented as a dryrun of manageUnits rather than being a
special deployment mode in order that plans don't interfere with the use
of deployment modes for operations like blue/green cutovers.