feat: "Account" and fixed build scope
Add support for a build scope of "account". This permits the same build
to be referenced across multiple segments/environments in an account. It
replaces the product identifier with the fixed value "account".

The initial use case for account scope is for code units that are shared
acros multiple products in an account.

Also support for a fixed value build scope. This permits all segments in
a product to reference the same build.

The initial use case for account scope is lambda authorisers. For single
product accounts, account and fixed scope would achieve the same result.
The choice is more to self-document the intention, i.e. across products
(account) or within a product (fixed).

The additional scopes have been added to RDS snapshots and docker
for consistency.