refactor: scope/formats from shared build (#96)
Refactor the script based management of builds to permit the formats
and scope for a build to be determined from a shared build file.

This will mean that build.json files only need to contain the commit
and optional tag, and formats/scopes need only be defined once as part
of the solution via the shared build directory tree.

The freemarker engine already supports this approach. A planned changes
is to move build updates to inside the freemarker engine so this logic
is contained in one place.