feat: Support shared build.json and shared_build.json
Both file can be present in the shared build tree. As scope/segment
refer to the build itself, it seems more logical to put those in a
build.json file in the shared tree, while if the unit is actually
shared, then a reference in a shared_build.json suggests the correct
semantic.