feat: Force all units to same commit
If APPLY_TO_ALL_DEPLOYMENT_UNITS is set to true, use the same git commit
for all provided units.

The use of stream jobs triggered from tags of code repos means that all
the information related to which unit was built from what repo is
effectively quarantined in the code repo Jenkinsfile.

The same commit/tag then needs to be applied to the build references of
all provided units. The APPLY_TO_ALL_DEPLOYMENT_UNITS setting does this
for tags, but if the commit is not set, the mapping between units and
code repos still needs to be known when managing build references.

To avoid the need to configure this information via the existing
*_CODE_REPO settings, extend the function of the APPLY_TO_ALL_DEPLOYMENT_UNITS
setting to also apply any git commit found to all provided units. The
existing logic currently does this by checking if the provided tag is in
fact a full commit id, but does not then permit the tag to be set as
well as the commit.

Tha assumption in all this is that the provided tag and commit are
consistent, which is the case given the way the stream jobs are triggered.