feat: add conventional commit formatting
Add utility routines to format the message of a git commit according to
the conventional commit (https://www.conventionalcommits.org/).

Also add support for local directory control to the git utility
routines so the working directory doesn't have to be within the repo,
and split the formatting of the git url from the other routines which
use it.