feat: custom entrance parameters (#120)
Add the ablity to pass arbitirary parameters to an entrance

Entrances can have quite disparate parameter needs. This PR adds the
ability to pass arbitrary parameters to an entrance, with the intention
a wrapper script will validate and collect these for use by the generate
template script. The entrance can then validate that the specific
parameters it needs have been provided.

The approach of wrapper scripts is already in use for blueprints and
buildBlueprints.