fix: limit conventional commit length
Limit the conventional commit message to 48 chars to conform with git
best practices.

Fix a bug in the use of clone when copying node content.

Ensure format_git_url gracefully handles the absence of credentials.

It now assumes no credentials means the repo is public and no http
authentication is inserted into the url.