refactor: remove cmdb tagging
The tagging of the cmdb to reflect automation activities has not
proven of use so this PR removes it from the various points where it
was previously performed.

Given that the same information as the tag is included in the cmdb
commit message, the PR refactors it to be compliant with the
conventional commit standard, which makes it more usable. To minimise
the change needed to achieve this, the commit message provided to
manageRepo is assumed to be name/value pairs and key named values used
to construct the conventional commit. Two extra named parameters are
included to allow explicit control over the type and description.