refactor: remove tagging of accounts cmdb
In line with https://github.com/hamlet-io/executor-bash/pull/129, this
change removes the tagging of changes made to the account cmdb.

Consistent with the approach to product configuration references, it
introduces an ACCOUNT_CONFIG_COMMIT to use as the confoguration references
value when updating accounts.

The commit messages will also start being formatted as conventional
commits, with the first account id as the scope.