feat: provider loading from contract
This implements provider sourcing and configuration from a CMDB via a
contract
The engine generates a loader contract which outlines the providers
which need to be loaded. Using this information the executor
establishes a provider cache and retrives the providers/plugins from
their sources.
The cache allows for the providers to be shared across different
districts when working locally

The process also generates a state file which can be used to determine
the version of the different hamlet providers that were used for a given
pass

If the setup script isn't used to load the providers then the existing
processes will continue working

Also adds some utility routines to handle urls for git auth