fix: git commi for builds
Ensure the git commit is captured in setContext.sh even if no deployment
units are known initially

A previous change (https://github.com/hamlet-io/executor-bash/pull/101)
to ensure the git commit was applied to all units if
desired did not account for the situation of builds, where the
deployment unit to be used is not known until later in the process, but
the git commit needs to be captured early.

This commit detects that condition and captures the git commit.