feat: account deployment unit region override
Support the ability to override the region an account level unit is
deployed into in the same manner as is currently supported for products.

Also correct use of unicode quotes by replacing with ASCII equivalent.

This has been introduced to support the ability to create account level
SES configuration for receiving emails in the regions in which it is
supported.