feat: non zip image url sourcing
updates the image sourcing from Url script to handle files which aren't
zip files. When this happens the hash of the URL image is still used but
before uploading the file into the registry the file is zipped up so
that we can treat it like our standard image registry files.

This is useful when you only have to provide one file and removes the
need to zip it up