feat: engine status codes
Refactor the template creation process to better capture and analyse
the output of the freemarker wrapper. Also return specific return
codes for `#stop` requests and for template processing errors.

By identifying these conditions, it is possible to integrate this with
more general automation to report template generation failures as soon
as they occur.