refactor: composite template inclusion
Include the contents of pre-assebled composite files rather than
relying on the `[#include]` directive to process their content.

As part of moving to dynamic input loading, reliance on pre-assembled
content needs to be removed. This change starts the process by using
the `?interpret` builtin internally to avoid use of the `[#include]`
directive.