fix: freemarker output redirection
Freemarker wrapper now relies on its environment to control where
stdout and stderr go.

Previously, the wrapper was always redirecting both stdout and stderr
to the one target. With the changes to output handling in the wrapper
and engine, the wrapper script now assumes the environment is set up
to redirect output where required.

Also remove the logging of "generation successful" to reduce the noise
in the output generated during template creation.