refactor: dynamic cmdb plugin detection (#218)
* refactor: dynamic cmdb plugin detection

Make the token used to detect use of the cmdb plugin more specific so
directory names containing `cmdb` don't result in a false positive.

* fix: Allow for directories containing "cmdb"

Broaden the pattern used for detecting the cmdb plugin by including
any path that ends with "cmdb". This supports the paths where the
docker container and bootstrap put the plugin.

The trailing semicolon ensures the detection will not trigger for paths
that contain "cmdb". A consequence is that if including the cmdb path
as the last entry in GENERATION_PLUGIN_DIRS, it must be followed by a
trailing semicolon. Note that this is not required if the directory uses
the plugin name i.e. "engine-plugin-cmdb".

* refactor: slightly more robust check for cmdb plugin