fix: resetting of temporary file stack
Ensure the temporary file stack is not reset when the `utility.sh is
sourced multiple times.

Previously each invocation of `utility.sh` would empty the temporary file
stack. With the adding of a call to `setCredentials.sh` (which sources
`utility.sh`) in createTemplate.sh when running a script, the
previously established temporary directory was lost resulting in the
failure of subsequent template passes.

Along the way, a couple of other things were noted and fixed;
- the cmdb migration process was still being run by `constructTree.sh`
- the saving of registry types was happening BEFORE `AUTOMATION_DATA_DIR`
was set, this variable being use by the `save_context_property` function.
As a result, the registry types were being saved to a temporary file
which was then not accessible once `AUTOMATION_DATA_DIR` was defined.